aws_ec2_prod/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcelo Trylesinski <marcelotryle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
pymongo = "^3.11.3"
dnspython = "^2.1.0"
requests = "^2.25.1"
stripe = "^2.56.0"
okta-jwt = "^1.3.5"
aiofiles = "^0.6.0"
Jinja2 = "^2.11.3"
numpy = "^1.20.2"
matplotlib = "^3.4.1"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
yapf = "^0.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
